Topic Deletion (#40)
* Created a simple setup for deleting topics.
Topic config files that include a top level entry of  will result in the topic being deleted.

There is currently no other check/balance on the topic. If the config file is changed to include that flag, that topic will be removed. I'd love to hear any feedback about other potential handles to make this safer.

Testing: Tested locally with a local kafka cluster that Ping helped me set up.

* Added some extra config to disable deletions on a cluster by cluster basis.

Co-authored-by: Justin Casol <jcasol@pinterest.com>